backend/requirements.txt
fastapi
uvicorn
networkx
pydantic
python-dotenv
google-generativeai
google-genai
pydantic-settings
